move declaration of chatSocket from app.js to loginUtils.js
